+++ comb_spec_searcher/specification.py
@@ -5,7 +5,7 @@ where each of the bi appear exactly once on the left hand side of some rule.
 from copy import copy
 from functools import reduce
 from operator import mul
-from typing import Dict, Generic, Iterable, Iterator, List, Sequence, Tuple
+from typing import Dict, Generic, Iterable, Iterator, List, Sequence, Tuple, Union
 
 import sympy
 from logzero import logger
@@ -51,9 +51,6 @@ class CombinatorialSpecification(
     A combinatorial specification is a set rules of the form a -> b1, ..., bk
     where each of the bi appear exactly once on the left hand side of some
     rule.
-
-    The default is to expand verified classes, but this can be turned off by
-    setting expand_verified to False.
     """
 
     def __init__(
@@ -66,16 +63,16 @@ class CombinatorialSpecification(
             ]
         ],
         equivalence_paths: Iterable[Sequence[CombinatorialClassType]],
-        expand_verified: bool = True,
     ):
         self.root = root
         self.rules_dict: Dict[CombinatorialClassType, AbstractRule] = {}
-        self._populate_rules_dict(strategies, equivalence_paths, expand_verified)
+        self._populate_rules_dict(strategies, equivalence_paths)
         for rule in list(
             self.rules_dict.values()
         ):  # list as we lazily assign empty rules
             rule.set_subrecs(self.get_rule)
         self.labels: Dict[CombinatorialClassType, int] = {}
+        self._label_to_tiling: Dict[int, CombinatorialClassType] = {}
 
     def _populate_rules_dict(
         self,
@@ -86,12 +83,10 @@ class CombinatorialSpecification(
             ]
         ],
         equivalence_paths: Iterable[Sequence[CombinatorialClassType]],
-        expand_verified: bool,
     ) -> None:
         equivalence_rules: Dict[
             Tuple[CombinatorialClassType, CombinatorialClassType], Rule
         ] = {}
-        verification_packs: Dict[CombinatorialClassType, StrategyPack] = {}
         for comb_class, strategy in strategies:
             if isinstance(strategy, AlreadyVerified):
                 continue
@@ -102,20 +97,44 @@ class CombinatorialSpecification(
                 equivalence_rules[(comb_class, non_empty_children[0])] = (
                     rule if len(rule.children) == 1 else rule.to_equivalence_rule()
                 )
-            elif non_empty_children:
+            elif non_empty_children or isinstance(rule, VerificationRule):
                 self.rules_dict[comb_class] = rule
-            elif isinstance(rule, VerificationRule):
-                if expand_verified:
-                    try:
-                        verification_packs[comb_class] = rule.pack()
-                    except InvalidOperationError:
-                        self.rules_dict[comb_class] = strategy(comb_class)
-                else:
-                    self.rules_dict[comb_class] = strategy(comb_class)
             else:
                 raise ValueError("Non verification rule has no children.")
         self._add_equivalence_path_rules(equivalence_paths, equivalence_rules)
+
+    def expand_verified(self) -> None:
+        """
+        Will expand all verified classes with respect to the strategy pack
+        given by the VerificationStrategies.
+        """
+        verification_packs: Dict[CombinatorialClassType, StrategyPack] = {}
+        for comb_class, rule in self.rules_dict.items():
+            if isinstance(rule, VerificationRule):
+                try:
+                    verification_packs[comb_class] = rule.pack()
+                except InvalidOperationError:
+                    logger.info("Can't expand the rule:\n%s", rule)
+        for comb_class in verification_packs:
+            self.rules_dict.pop(comb_class)
         self._expand_verified_comb_classes(verification_packs)
+
+    def expand_comb_class(self, comb_class: Union[int, CombinatorialClassType]) -> None:
+        """
+        Will try to expand a particular class with respect to the strategy pack
+        that the VerificationStrategy has.
+        """
+        rule = self.get_rule(comb_class)
+        if isinstance(rule, VerificationRule):
+            try:
+                pack = rule.pack()
+            except InvalidOperationError:
+                logger.info("Can't expand the rule:\n%s", rule)
+                return
+            self.rules_dict.pop(rule.comb_class)
+            self._expand_verified_comb_classes({rule.comb_class: pack})
+        else:
+            logger.info("Can't expand the rule:\n%s", rule)
 
     def _add_equivalence_path_rules(
         self,
@@ -152,12 +171,20 @@ class CombinatorialSpecification(
             logger.info(css.run_information())
             # pylint: disable=protected-access
             comb_class_rules, comb_class_eqv_paths = css._auto_search_rules()
-            self._populate_rules_dict(comb_class_rules, comb_class_eqv_paths, True)
+            self._populate_rules_dict(comb_class_rules, comb_class_eqv_paths)
 
     def get_rule(
-        self, comb_class: CombinatorialClassType
+        self, comb_class: Union[int, CombinatorialClassType]
     ) -> AbstractRule[CombinatorialClassType, CombinatorialObjectType]:
         """Return the rule with comb class on the left."""
+        if isinstance(comb_class, int):
+            try:
+                comb_class = self._label_to_tiling[comb_class]
+            except KeyError:
+                raise KeyError(
+                    f"The label {comb_class} does not correspond to a tiling"
+                    " in the specification."
+                )
         if comb_class not in self.rules_dict:
             if comb_class.is_empty():
                 empty_strat = EmptyStrategy()
@@ -177,6 +204,7 @@ class CombinatorialSpecification(
         if res is None:
             res = len(self.labels)
             self.labels[comb_class] = res
+            self._label_to_tiling[res] = comb_class
         return res
 
     def get_function(self, comb_class: CombinatorialClassType) -> Function:
@@ -432,7 +460,7 @@ class CombinatorialSpecification(
         }
 
     @classmethod
-    def from_dict(cls, d, expand_verified=False):
+    def from_dict(cls, d):
         """
         Return the specification with the dictionary outputter by the
         'to_jsonable' method
@@ -450,7 +478,6 @@ class CombinatorialSpecification(
                 [CombinatorialClass.from_dict(class_dict) for class_dict in eqv_path]
                 for eqv_path in d["eqv_paths"]
             ],
-            expand_verified=expand_verified,
         )